+++ index.ts
@@ -10,7 +10,7 @@ import ComplexData from './src/data/ComplexData'
 
 import ArrayValue from './src/lib/ArrayValue'
 import AttrsValue from './src/lib/AttrsValue'
-import CascadeValue from './src/lib/CascadeValue'
+import CascaderValue from './src/lib/CascaderValue'
 import DictionaryValue from './src/lib/DictionaryValue'
 import FormValue from './src/lib/FormValue'
 import InterfaceValue from './src/lib/InterfaceValue'
@@ -56,7 +56,7 @@ export {
 
   ArrayValue,
   AttrsValue,
-  CascadeValue,
+  CascaderValue,
   DictionaryValue,
   FormValue,
   InterfaceValue,